+++ src/client/index.ts
@@ -1,15 +1,15 @@
 import {
   APIClientOptions,
-  RumbleAccount,
-  RumbleAsset,
-  RumbleOrganization,
-  RumbleSite,
-  RumbleUser,
+  RunZeroAccount,
+  RunZeroAsset,
+  RunZeroOrganization,
+  RunZeroSite,
+  RunZeroUser,
 } from '../types';
 import { AccountAPIKeyClient } from './APIKeyClient';
 import { ExportTokenApiClient } from './ExportTokenClient';
 
-export const BASE_URI = 'https://console.rumble.run/api/v1.0';
+export const BASE_URI = 'https://console.runzero.com/api/v1.0';
 export type ResourceIteratee<T> = (each: T) => Promise<void> | void;
 
 export function createAPIClient(options: APIClientOptions): APIClient {
@@ -22,11 +22,11 @@ export function createAPIClient(options: APIClientOptions): APIClient {
 
 export interface APIClient {
   verifyAuthentication(): Promise<void>;
-  getAccount(): Promise<RumbleAccount> | RumbleAccount;
+  getAccount(): Promise<RunZeroAccount> | RunZeroAccount;
   iterateOrganizations(
-    iteratee: ResourceIteratee<RumbleOrganization>,
+    iteratee: ResourceIteratee<RunZeroOrganization>,
   ): Promise<void>;
-  iterateUsers(iteratee: ResourceIteratee<RumbleUser>): Promise<void>;
-  iterateSites(iteratee: ResourceIteratee<RumbleSite>): Promise<void>;
-  iterateAssets(iteratee: ResourceIteratee<RumbleAsset>): Promise<void>;
+  iterateUsers(iteratee: ResourceIteratee<RunZeroUser>): Promise<void>;
+  iterateSites(iteratee: ResourceIteratee<RunZeroSite>): Promise<void>;
+  iterateAssets(iteratee: ResourceIteratee<RunZeroAsset>): Promise<void>;
 }