+++ src/steps/account/index.test.ts
@@ -4,7 +4,7 @@ import {
 } from '../../../test/config';
 import {
   createStepCollectionTest,
-  rumbleRecordingOptions,
+  runZeroRecordingOptions,
 } from '../../../test/recording';
 import { Steps } from '../constants';
 
@@ -15,7 +15,7 @@ describe('#fetchAccountDetails - API Key', () => {
       recordingSetup: {
         directory: __dirname,
         name: 'fetchAccountDetailsShouldCollectData',
-        ...rumbleRecordingOptions,
+        ...runZeroRecordingOptions,
       },
       stepConfig: buildStepTestConfigForAPIKey(Steps.ACCOUNT),
     }),
@@ -24,14 +24,14 @@ describe('#fetchAccountDetails - API Key', () => {
 
 describe('#fetchAccountDetails - Export Tokens', () => {
   // NOTE: This test won't produce a recording under current implementation
-  // of client. Rumble account produced using static integration data.
+  // of client. runZero account produced using static integration data.
   test(
     'should collect data and create account entity',
     createStepCollectionTest({
       recordingSetup: {
         directory: __dirname,
         name: 'fetchAccountDetailsExportTokens',
-        ...rumbleRecordingOptions,
+        ...runZeroRecordingOptions,
       },
       stepConfig: buildStepTestConfigForExportToken(Steps.ACCOUNT),
     }),