+++ src/steps/asset/converter.ts
@@ -3,10 +3,10 @@ import {
   Entity,
   parseTimePropertyValue,
 } from '@jupiterone/integration-sdk-core';
-import { RumbleAsset } from '../../types';
+import { RunZeroAsset } from '../../types';
 import { Entities } from '../constants';
 
-export function createAssetEntity(asset: RumbleAsset): Entity {
+export function createAssetEntity(asset: RunZeroAsset): Entity {
   const name = selectNameForAsset(asset);
 
   return createIntegrationEntity({
@@ -39,12 +39,21 @@ export function createAssetEntity(asset: RumbleAsset): Entity {
         osVersion: asset.os_version,
         macAddresses: asset.macs,
         deviceId: null,
+        serviceCount: asset.service_count,
+        softwareCount: asset.software_count,
+        vulnerabilityCount: asset.vulnerability_count,
+        lastAgentId: asset.last_agent_id,
+        lastTaskId: asset.last_task_id,
+        newestMacAddress: asset.newest_mac,
+        newestMacVendor: asset.newest_mac_vendor,
+        orgName: asset.org_name,
+        siteName: asset.site_name,
       },
     },
   });
 }
 
-function selectNameForAsset(asset: RumbleAsset) {
+function selectNameForAsset(asset: RunZeroAsset) {
   if (asset.names && asset.names.length > 0) {
     return asset.names[0];
   } else if (asset.hw) {